重构: 把 deprecated resolvePluginRuntimeOwner 移到 test fixture (PR#487 review #9)
回应原作者 review #9：deprecated 函数仍被测试 import，但留在 src 里有
被生产代码再次误引用的风险。

- 新建 tests/helpers/legacy-runtime-owner.ts 持有函数体（仅作 fixture）
- 删除 src/plugin-command-expander.ts 中的 @deprecated 段（按语义边界
  止于 // --- Plugin runtime path resolution --- 注释，下一段
  resolvePluginRoot 保留）
- src/runtime-owner.ts:63 与 src/index.ts:5822 的注释引用一并改写为
  「mirroring the legacy virtual-JID base check」，避免再提 deprecated
  helper 名
- tests/plugin-expander-routing-bugs.test.ts 的 import 切到新 helper
- rg -n "resolvePluginRuntimeOwner" src 输出为空验证通过
